Major Balance and pacing update

+++ Catching waves/Main/Rythmic_Objects.pde
@@ -1,7 +1,7 @@
 public class Rythmic_Objects{
     PImage image;
-    int x = 1000;
-    int y = 32;
+    float x = 1000;
+    float y = 32;
     int hor = 64,ver = 64;
     int correctInput;
